expose some headers
