+++ rmw_connext_cpp/include/rmw_connext_cpp/connext_static_publisher_info.hpp
@@ -12,22 +12,22 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef TYPES__CONNEXT_STATIC_CLIENT_INFO_HPP_
-#define TYPES__CONNEXT_STATIC_CLIENT_INFO_HPP_
+#ifndef RMW_CONNEXT_CPP__CONNEXT_STATIC_PUBLISHER_INFO_HPP_
+#define RMW_CONNEXT_CPP__CONNEXT_STATIC_PUBLISHER_INFO_HPP_
 
-#include "rmw_connext_shared_cpp/ndds_include.hpp"
+#include "rmw_connext_shared_cpp/types.hpp"
 
-#include "rosidl_typesupport_connext_cpp/service_type_support.h"
+#include "rosidl_typesupport_connext_cpp/message_type_support.h"
 
 extern "C"
 {
-struct ConnextStaticClientInfo
+struct ConnextStaticPublisherInfo
 {
-  void * requester_;
-  DDSDataReader * response_datareader_;
-  DDSReadCondition * read_condition_;
-  const service_type_support_callbacks_t * callbacks_;
+  DDSPublisher * dds_publisher_;
+  DDSDataWriter * topic_writer_;
+  const message_type_support_callbacks_t * callbacks_;
+  rmw_gid_t publisher_gid;
 };
 }  // extern "C"
 
-#endif  // TYPES__CONNEXT_STATIC_CLIENT_INFO_HPP_
+#endif  // RMW_CONNEXT_CPP__CONNEXT_STATIC_PUBLISHER_INFO_HPP_
