+++ rmw_connext_cpp/include/rmw_connext_cpp/connext_static_service_info.hpp
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef TYPES__CONNEXT_STATIC_SERVICE_INFO_HPP_
-#define TYPES__CONNEXT_STATIC_SERVICE_INFO_HPP_
+#ifndef RMW_CONNEXT_CPP__CONNEXT_STATIC_SERVICE_INFO_HPP_
+#define RMW_CONNEXT_CPP__CONNEXT_STATIC_SERVICE_INFO_HPP_
 
 #include "rmw_connext_shared_cpp/ndds_include.hpp"
 
@@ -30,4 +30,4 @@ struct ConnextStaticServiceInfo
 };
 }  // extern "C"
 
-#endif  // TYPES__CONNEXT_STATIC_SERVICE_INFO_HPP_
+#endif  // RMW_CONNEXT_CPP__CONNEXT_STATIC_SERVICE_INFO_HPP_
