+++ rmw_connext_cpp/include/rmw_connext_cpp/connext_static_subscriber_info.hpp
@@ -12,13 +12,23 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef IDENTIFIER_HPP_
-#define IDENTIFIER_HPP_
+#ifndef RMW_CONNEXT_CPP__CONNEXT_STATIC_SUBSCRIBER_INFO_HPP_
+#define RMW_CONNEXT_CPP__CONNEXT_STATIC_SUBSCRIBER_INFO_HPP_
+
+#include "rmw_connext_shared_cpp/ndds_include.hpp"
+
+#include "rosidl_typesupport_connext_cpp/message_type_support.h"
 
 extern "C"
 {
-// static for internal linkage
-extern const char * const rti_connext_identifier;
+struct ConnextStaticSubscriberInfo
+{
+  DDSSubscriber * dds_subscriber_;
+  DDSDataReader * topic_reader_;
+  DDSReadCondition * read_condition_;
+  bool ignore_local_publications;
+  const message_type_support_callbacks_t * callbacks_;
+};
 }  // extern "C"
 
-#endif  // IDENTIFIER_HPP_
+#endif  // RMW_CONNEXT_CPP__CONNEXT_STATIC_SUBSCRIBER_INFO_HPP_
