+++ rmw_connext_cpp/include/rmw_connext_cpp/identifier.hpp
@@ -12,23 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef TYPES__CONNEXT_STATIC_SUBSCRIBER_INFO_HPP_
-#define TYPES__CONNEXT_STATIC_SUBSCRIBER_INFO_HPP_
+#ifndef RMW_CONNEXT_CPP__IDENTIFIER_HPP_
+#define RMW_CONNEXT_CPP__IDENTIFIER_HPP_
 
-#include "rmw_connext_shared_cpp/ndds_include.hpp"
+extern const char * const rti_connext_identifier;
 
-#include "rosidl_typesupport_connext_cpp/message_type_support.h"
-
-extern "C"
-{
-struct ConnextStaticSubscriberInfo
-{
-  DDSSubscriber * dds_subscriber_;
-  DDSDataReader * topic_reader_;
-  DDSReadCondition * read_condition_;
-  bool ignore_local_publications;
-  const message_type_support_callbacks_t * callbacks_;
-};
-}  // extern "C"
-
-#endif  // TYPES__CONNEXT_STATIC_SUBSCRIBER_INFO_HPP_
+#endif  // RMW_CONNEXT_CPP__IDENTIFIER_HPP_
