+++ rmw_connext_cpp/src/rmw_compare_gid_equals.cpp
@@ -18,7 +18,7 @@
 
 #include "rmw_connext_shared_cpp/types.hpp"
 
-#include "identifier.hpp"
+#include "rmw_connext_cpp/identifier.hpp"
 
 extern "C"
 {
