+++ rmw_connext_cpp/src/rmw_count.cpp
@@ -16,7 +16,7 @@
 
 #include "rmw_connext_shared_cpp/count.hpp"
 
-#include "identifier.hpp"
+#include "rmw_connext_cpp/identifier.hpp"
 
 extern "C"
 {
