+++ rmw_connext_cpp/src/rmw_get_gid_for_publisher.cpp
@@ -16,8 +16,8 @@
 #include "rmw/impl/cpp/macros.hpp"
 #include "rmw/rmw.h"
 
-#include "identifier.hpp"
-#include "types/connext_static_publisher_info.hpp"
+#include "rmw_connext_cpp/identifier.hpp"
+#include "rmw_connext_cpp/connext_static_publisher_info.hpp"
 
 extern "C"
 {
