+++ rmw_connext_cpp/src/rmw_node_names.cpp
@@ -18,7 +18,7 @@
 
 #include "rmw_connext_shared_cpp/node_names.hpp"
 
-#include "identifier.hpp"
+#include "rmw_connext_cpp/identifier.hpp"
 
 extern "C"
 {
