+++ rmw_connext_cpp/src/rmw_service_server_is_available.cpp
@@ -20,8 +20,8 @@
 #include "rmw_connext_shared_cpp/count.hpp"
 #include "rmw_connext_shared_cpp/ndds_include.hpp"
 
-#include "identifier.hpp"
-#include "types/connext_static_client_info.hpp"
+#include "rmw_connext_cpp/identifier.hpp"
+#include "rmw_connext_cpp/connext_static_client_info.hpp"
 
 // Uncomment this to get extra console output about discovery.
 // This affects code in this file, but there is a similar variable in:
