+++ rmw_connext_cpp/src/rmw_topic_names_and_types.cpp
@@ -17,7 +17,7 @@
 
 #include "rmw_connext_shared_cpp/topic_names_and_types.hpp"
 
-#include "identifier.hpp"
+#include "rmw_connext_cpp/identifier.hpp"
 
 extern "C"
 {
