+++ rmw_connext_cpp/src/rmw_trigger_guard_condition.cpp
@@ -16,7 +16,7 @@
 
 #include "rmw_connext_shared_cpp/trigger_guard_condition.hpp"
 
-#include "identifier.hpp"
+#include "rmw_connext_cpp/identifier.hpp"
 
 extern "C"
 {
