+++ rmw_connext_cpp/src/rmw_wait.cpp
@@ -16,10 +16,10 @@
 
 #include "rmw_connext_shared_cpp/wait.hpp"
 
-#include "identifier.hpp"
-#include "types/connext_static_client_info.hpp"
-#include "types/connext_static_service_info.hpp"
-#include "types/connext_static_subscriber_info.hpp"
+#include "rmw_connext_cpp/identifier.hpp"
+#include "rmw_connext_cpp/connext_static_client_info.hpp"
+#include "rmw_connext_cpp/connext_static_service_info.hpp"
+#include "rmw_connext_cpp/connext_static_subscriber_info.hpp"
 
 extern "C"
 {
