+++ rmw_connext_cpp/src/rmw_waitset.cpp
@@ -16,7 +16,7 @@
 
 #include "rmw_connext_shared_cpp/waitset.hpp"
 
-#include "identifier.hpp"
+#include "rmw_connext_cpp/identifier.hpp"
 
 extern "C"
 {
